Add urlify function to utils.ts
